+++ examples/bulma.py
@@ -680,7 +680,6 @@ if __name__=="__main__":
     logging.getLogger("htag.tag").setLevel( logging.WARNING )
 
     from htag.runners import *
-    # r=GuyApp( Page )
     # r=PyWebWiew( Page )
     # r=BrowserStarletteHTTP( Page )
     # r=BrowserStarletteWS( Page )