+++ test_runners.py
@@ -18,7 +18,6 @@ def data_source():
         "BrowserStarletteWS",
         "WebHTTP",
         "PyWebWiew",
-        "GuyApp",
         "ChromeApp",
         "AndroidApp",
         "BrowserTornadoHTTP",
@@ -28,12 +27,16 @@ def data_source():
 
 @pytest.mark.parametrize('my_runner', data_source())
 def test_a_runner( my_runner ):
-
+    # all external lib for runners, should be here, to 100% ok for this test !
+    # all external lib for runners, should be here, to 100% ok for this test !
+    # all external lib for runners, should be here, to 100% ok for this test !
+    # all external lib for runners, should be here, to 100% ok for this test !
     mrunners=importlib.import_module("htag.runners")
 
     if hasattr(mrunners,my_runner):
         runner=getattr(mrunners,my_runner)
-        runner( MyApp )
+        r=runner( MyApp )
+        assert hasattr(r,"run")
     else:
         print("can't test %s" % my_runner)